fixed warnings for signed/unsigned conversion, shortening, etc (fixed by using the convert_integer<A,B>() function)

+++ asdf_multiplat/data/settings.h
@@ -4,8 +4,8 @@ namespace asdf {
 
     struct settings_t {
 
-        size_t resolution_width = 1280;
-        size_t resolution_height = 720;
+        uint32_t resolution_width = 1280;
+        uint32_t resolution_height = 720;
 
         bool fullscreen = false;
         bool borderless = false;
